[FIXED] 修复了一些已知的bug
[UPDATE]
新增了学生和家长以及班级信息获取的接口，并已通过测试

[FINISH]
项目当前所有的功能模块的编码工作均完成，并且目前已完成了所有编写的接口的单元测试，后续将继续根据需求进行一些架构或代码的优化

+++ data-access/src/main/java/sc/hqu/graduationdesign/homeworkmanager/model/ClassStudentParentView.java
@@ -13,7 +13,7 @@ public class ClassStudentParentView {
     /**
      * 班级id
      */
-    private Integer classId;
+    private Long classId;
 
     /**
      * 学号
@@ -28,7 +28,7 @@ public class ClassStudentParentView {
     /**
      * 家长id
      */
-    private Integer parentId;
+    private Long parentId;
 
     /**
      * 家长姓名
